feat(firebase): extract firebase import
