+++ src/components/Admin/Bulk/components/Filter.vue
@@ -63,7 +63,7 @@
     </md-card>
 </template>
 <script>
-import db from '@/database';
+import db from '@/firebase/database';
 ///import fetchUserDatas from '@/methods/fetchUserDatas.js';
 
 export default {
