+++ src/components/Admin/Bulk/components/Results.vue
@@ -111,7 +111,7 @@
     </div>
 </template>
 <script>
-import db from '@/database';
+import db from '@/firebase/database';
 
 export default {
     name: 'BulkEdit-Results',
