+++ src/components/Admin/Export/App.vue
@@ -42,7 +42,7 @@
     </div>
 </template>
 <script>
-import db from '@/database';
+import db from '@/firebase/database';
 
 export default {
     name: "Export",
