+++ src/components/Exam/App.vue
@@ -16,7 +16,7 @@
     </div>
 </template>
 <script>
-import db from '@/database';
+import db from '@/firebase/database';
 
 export default {
     name: 'ExamPage',
