+++ src/components/Profile/App.vue
@@ -65,7 +65,7 @@
     </div>
 </template>
 <script>
-import db from '@/database';
+import db from '@/firebase/database';
 
 import QuestionList from './QuestionList.vue'
