+++ src/components/Topic/Form.vue
@@ -84,7 +84,7 @@ import * as firebase from 'firebase/app';
 import 'firebase/firestore';
 import "firebase/auth";
 
-import db from '@/database';
+import db from '@/firebase/database';
 
 export default {
     name: 'Send',
