+++ src/components/Topic/components/NewQuestions.vue
@@ -23,7 +23,7 @@
     </div>
 </template>
 <script>
-import db from '@/database';
+import db from '@/firebase/database';
 
 import QuestionCard from './QuestionCard.vue';
 import fetchUserDatas from '@/methods/fetchUserDatas.js';
