+++ src/firebase/auth/index.js
@@ -1,0 +1,6 @@
+import firebase from '../';
+import 'firebase/auth';
+
+const auth = firebase.auth();
+
+export default auth;
