+++ src/firebase/index.js
@@ -1,0 +1,18 @@
+/*
+	Import & Load Firebase base
+*/
+import * as firebase from 'firebase/app';
+
+/*
+	Initialize Firebase App using environment variables
+*/
+
+firebase.initializeApp({
+	apiKey: process.env.VUE_APP_FIREBASE_API_KEY,
+	authDomain: process.env.VUE_APP_FIREBASE_AUTH_DOMAIN,
+	databaseURL: process.env.VUE_APP_FIREBASE_DATABASE_URL,
+	projectId: process.env.VUE_APP_FIREBASE_PROJECT_ID,
+	storageBucket: process.env.VUE_APP_FIREBASE_STORAGE_BUCKET,
+});
+
+export default firebase;
