+++ src/main.js
@@ -69,11 +69,6 @@ Vue.use(MdContent)
 Vue.config.productionTip = false;
 
 /*
-	Import & Load Firebase base
-*/
-import * as firebase from 'firebase/app';
-
-/*
 	Import & Load VueFire
 */
 import VueFire from 'vuefire';
@@ -136,17 +131,6 @@ if (process.env.NODE_ENV == 'production' && process.env.VUE_APP_SENTRY_API) {
 */
 import VuePaginate from 'vue-paginate'
 Vue.use(VuePaginate)
-
-/*
-	Initialize Firebase App using environment variables
-*/
-firebase.initializeApp({
-	apiKey: process.env.VUE_APP_FIREBASE_API_KEY,
-	authDomain: process.env.VUE_APP_FIREBASE_AUTH_DOMAIN,
-	databaseURL: process.env.VUE_APP_FIREBASE_DATABASE_URL,
-	projectId: process.env.VUE_APP_FIREBASE_PROJECT_ID,
-	storageBucket: process.env.VUE_APP_FIREBASE_STORAGE_BUCKET,
-});
 
 /* eslint-disable no-new */
 new Vue({
